openssl: delete support for OCSP stapling
OCSP has been obsolete for a while now. I thought OCSP Must-Staple was a
pretty good idea, but it has failed and there is no point to supporting
it anymore. Time to give up and move on.

GnuTLS users will still actually receive OCSP verification so long as
GnuTLS continues to do so by default, but this is probably a bad thing
if anything, because no web browsers do so, and occasionally users
encounter prominent websites that present invalid OCSP responses. I
don't think maintaining the same behavior for the OpenSSL backend is
useful.

Part-of: <https://gitlab.gnome.org/GNOME/glib-networking/-/merge_requests/265>